feat: ability to override a way to obtain a file handle
Hello! Thanks for the great library!

I tried to use it for reading meta-data from various Blender files and some of them are compressed. Thus, I want to provide different file handlers (such as gzip reader) for the `__init__`.  So just extracted `_get_file_handler` method which can be overridden in inherited classes.